requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
redis==5.0.0
celery==5.3.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
zipfile36==0.1.3